refactor: 관제 페이지 refactor
