+++ components/video/index.ts
@@ -1,8 +1,13 @@
 export { JoinBanner } from './JoinBanner';
 export { ControlBar } from './ControlBar';
+export { ControlBarWrapper } from './ControlBarWrapper';
 export { ParticipantSidebar } from './ParticipantSidebar';
 export { ParticipantDetailSidebar } from './ParticipantDetailSidebar';
 export { FullScreenVideo } from './FullScreenVideo';
+export { EmptyState } from './EmptyState';
+export { RoomTracks } from './RoomTracks';
+export { TakeoverAudioController } from './TakeoverAudioController';
+export { LiveTileOps } from './LiveTileOps';
 export type { MockParticipant } from './ParticipantSidebar';
 export {
   EmptyTile,
